requirements.txt
python-dotenv
mcp
requests
psutil
fastmcp
uvicorn